+++ crates/aptos/src/update/mod.rs
@@ -1,7 +1,7 @@
 // Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-mod helpers;
+pub mod helpers;
 mod tool;
 
 use helpers::check_if_update_required;